Use minimal index.js
The index.js is required so the importer finds our module. Otherwise we
do not use it as kpop is always imported as es6 module. This change only
exposes the build version in the index and removes everything else where
it makes no sense and stuff instead should be directly imported.