Add TopBar component
All apps need a top bar. This adds a flexible component which by default
shows the Kopano logo with a title and offers options for an anchor on
the left (to show hide side panel), centered sub elements and right side
elements. It also can show a users profile when an appropriate user prop
is passed.