base: Add config context for config delegation
To allow nested components to utilize values from the app configuration,
this change adds a config context to the base container. If set by the
app, nested components can apply config values for their rendering.

As a first application for this new functionality, the app switcher uses
a config array value from `apps.enabled` to control visibility and
availablity of kopano apps. If this value is set to empty, also the app
switcher button and menu entry will not render at all.