oidc: Add support for sign in and sign out via popup
Inside progressive web app contexts it is not possible to redirect the
app to the external sign in and sign out pages without exiting the app.
This change adds popup based sign in and sign out support to kpop. It is
automatically turned on when kpop detects that the app is in standalone
mode.

Standalone mode detections works with standards complient css check
(works in Chrome), and with non-standard standalone attribute for
Safari.

To allow popup mode, the base application flows have been extended with
an intermediate sign in screen which offers a button to trigger/open the
external sign in popup window withing a user action.