oidc: Disable popup sign in mode in Mobile Safari
Mobile Safari reloads PWAs whenever they become visible. This makes this
environment unable to sign in. This change disables popup mode when in
standalone mode of Mobile Safari for the time being. Eventually Apple
might fix. This means when adding an App to the home screen on iOS the
app opens fine but then opens Safari. That at least makes the user flow
functional where otherwise a user was unable to sign in inside the
standalone app.

Reference: https://medium.com/@firt/pwas-are-coming-to-ios-11-3-cupertino-we-have-a-problem-2ff49fd7d6ea
Reference: https://forums.developer.apple.com/thread/100407