Support extra query params with oidc authorize
To allow clients to add additional parameters to the oidc authorize
request, this change adds the `oidc.eqp` config key. When set to an
object, its key/values will be used as additional string parameters to
the oidc authorize request. This functionality is required for certain
OIDC providers or to request additional behavior.