Avoid flash of sign-in screen when local user data is expired
The OIDC client library caches user data in sessionStorage. Then the app
is closed but the browser session remains, this sessionStorage data will
eventually expire. When opening the app again, this state is now checked
and instead of triggering the user required error, the app tries to
refresh the user first, resulting in better user experience by not
flashing the sign-in screen until the user has been refreshed in the
background.