Force OIDC mode to use fragment
Kpop apps are web apps and thus can always use fragment mode for OIDC
callback request parameters. This makes sure that the values are not
leaked to the server and as a side effect fixes a problem where the oidc
client library used by kpop does not correctly use fragment mode when
not explicitly told to do so.

This change for example fixes the popup callback when the OIDC server
does return query response parameters by default.