Implement refresh token support
To stay signed in, a refresh token can be requested from OIDC via the
offline_access scope. This change auto adds the offline_access scope to
the requested scopes if the oidc.standaloneOfflineAccess and the app is
running in standalone (installed PWA) mode. All the required gear is
added to remain signed in in that case, leveraging local storage to
persistently store the token data on the device.